refactor(constants): move `_constants.ts` -> `_test/constants.ts`
