+++ fs/findNearestFile.test.ts
@@ -1,6 +1,6 @@
 import { resolve } from "../_deps/path.ts";
 import { assertEquals, assertRejects, describe, it } from "../_deps/testing.ts";
-import { FIXTURE_DIR } from "../_constants.ts";
+import { FIXTURE_DIR } from "../_test/constants.ts";
 import { findNearestFile } from "./findNearestFile.ts";
 
 const A_DIR = resolve(FIXTURE_DIR, "fs", "a");
