+++ fs/glob.test.ts
@@ -1,7 +1,7 @@
 import { resolve } from "../_deps/path.ts";
 import { assertArrayIncludes, describe, it } from "../_deps/testing.ts";
 import { glob } from "./glob.ts";
-import { FIXTURE_DIR } from "../_constants.ts";
+import { FIXTURE_DIR } from "../_test/constants.ts";
 
 const GLOB = resolve(FIXTURE_DIR, "fs", "**", "*.ts");
